+++ scripts/editor.js
@@ -1,28 +1,7 @@
-export class VariachelsJournalSheet extends JournalTextPageSheet {
-    activateEditor(name, options = {}, initialContent = '') {
-        const hasButton = button && button.classList.contains('editor-edit')
-        options.plugins = _configureProseMirrorPlugins(name, hasButton)
-
-        // {
-        //     menu: APSJMenu.build(ProseMirror.defaultSchema, {
-        //         onSave: () => this.saveEditor(name, { remove: false }),
-        //     }),
-        // }
-        return super.activateEditor(name, options, initialContent)
-    }
-
-    _configureProseMirrorPlugins(name, { remove = true } = {}) {
-        return {
-            menu: ProseMirror.ProseMirrorMenu.build(ProseMirror.defaultSchema, {
-                destroyOnSave: remove,
-                onSave: () => this.saveEditor(name, { remove }),
-            }),
-            keyMaps: ProseMirror.ProseMirrorKeyMaps.build(
-                ProseMirror.defaultSchema,
-                {
-                    onSave: () => this.saveEditor(name, { remove }),
-                }
-            ),
-        }
+export class VariachelsJournalSheet extends JournalSheet {
+    _getPageData() {
+        const _pageData = super._getPageData()
+        if (_pageData[0]?.number === 0) _pageData.forEach((p) => p.number++)
+        return _pageData
     }
 }